Move RequiredArg, {variant,trigger}_preprocessing to typeconverter.py
